Added Essential Directories/Files
I've added the folders for screens, components etc. and moved all the code segments to their respective files.
