+++ package.json
@@ -18,7 +18,8 @@
     "react": "18.2.0",
     "react-native": "0.74.1",
     "react-native-safe-area-context": "^4.10.1",
-    "react-native-screens": "^3.31.1"
+    "react-native-screens": "^3.31.1",
+    "react-native-vector-icons": "^10.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
