+++ src/styles/colors.js
@@ -1,0 +1,5 @@
+export default {
+  primary: '#3930B3',
+  secondary: '#E8E8E8',
+  // more colors here according to our need
+};
